+++ docker/start.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+service php5-fpm start && /usr/sbin/nginx -g "daemon off;" "$@"